leam-base.commons-lang.mid-1140.idx-44426.21.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableLong.java
@@ -124,7 +124,7 @@ public class MutableLong extends Number implements Comparable<MutableLong>, Muta
     public long getAndIncrement() {
         final long last = value;
         value++;
-        return last;
+        return (long) last;
     }
 
     /**
